Moved sevrer call and list making of cites to a class.
